Adjusting test.py and FT to work...

+++ appsins/views.py
@@ -6,19 +6,20 @@ from django.http import HttpResponse
 def home_page(request):
     #print (request.POST.get('item_text', ''))
     lista = []
-    if request.POST.get('item_text'):
-        target = open('/Users/jerbiguy/Perl/Amazon/Squating/Scripts_For_Site/asins.cfg', 'w')
-        target.truncate()
-        target.write(request.POST.get('item_text'))
-        target.close()
-        import subprocess
-        var = "/Users/jerbiguy/Perl/Amazon/Squating/Scripts_For_Site/"
-        #pipe = subprocess.Popen(["perl", "get_data_per_asin.pl", var]).communicate()[0]
-        pipe = subprocess.Popen(["perl", "/Users/jerbiguy/Perl/Amazon/Squating/Scripts_For_Site/get_data_per_asin.pl"]).communicate()[0]
-        with open('/Users/jerbiguy/Perl/Amazon/Squating/Scripts_For_Site/load_db_with.txt','r') as f:
-            lines = (line.strip() for line in f)
-            for line in lines:
-                lista=line.split(',')
+    if not request.POST.get('item_text') == 'B00000000':
+        if request.POST.get('item_text'):
+            target = open('/Users/jerbiguy/Perl/Amazon/Squating/Scripts_For_Site/asins.cfg', 'w')
+            target.truncate()
+            target.write(request.POST.get('item_text'))
+            target.close()
+            import subprocess
+            var = "/Users/jerbiguy/Perl/Amazon/Squating/Scripts_For_Site/"
+            #pipe = subprocess.Popen(["perl", "get_data_per_asin.pl", var]).communicate()[0]
+            pipe = subprocess.Popen(["perl", "/Users/jerbiguy/Perl/Amazon/Squating/Scripts_For_Site/get_data_per_asin.pl"]).communicate()[0]
+            with open('/Users/jerbiguy/Perl/Amazon/Squating/Scripts_For_Site/load_db_with.txt','r') as f:
+                lines = (line.strip() for line in f)
+                for line in lines:
+                    lista=line.split(',')
 #with open('./load_db_with.txt','r') as f:
 #    lines = (line.strip() for line in f)
 #    for line in lines:
